Abstract parser logic in Parser class
